Modify docker base image to aiida-core (#55)
This PR changes the base image in Docker file to be aiida-core. The
Docker file itself is also a bit restructured: the RASPA and aiida-raspa
installation from now on does not happen in the user's home folder,
but in the `/opt` folder instead. The `KILL_ALL_RPOCESSES_TIMEOUT`
is set to 50 seconds.

Travis file is also changed (since tests are run in Docker image). The
`sleep 30` call is replaced by `wait-for-services` which releases the
command line once all startup scripts are run. At the end of the
execution of the tests, the docker log is printed.